requirements.txt
fastapi==0.115.6
uvicorn==0.20.0
pymongo==4.10.1
neo4j==5.17.0
pydantic==2.6.1
polyline==2.0.2
geojson==3.1.0
python-dotenv==1.0.1
pydantic_settings==2.1.0
requests==2.32.3
PyJWT==2.9.0
